requirements.txt
# Core pipeline dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0

# =========================================================================
# Deep Learning Classifier Dependencies (CausalGAT)
# =========================================================================
# Note: These are required for the DL classifier to work.
# If not installed, the classifier will fall back to placeholder mode.

# PyTorch (install separately based on your CUDA version)
# See: https://pytorch.org/get-started/locally/
# Example for CPU: pip install torch torchvision torchaudio
# Example for CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0

# PyTorch Geometric (for GNN operations)
# Install after PyTorch: pip install torch_geometric torch_scatter
torch_geometric>=2.4.0
torch_scatter>=2.1.0

# Scientific computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Optional: For code embeddings (when user provides implementation)
# transformers>=4.30.0  # For CodeT5
# node2vec>=0.4.0       # For graph embeddings
